*refactoring of ConvertPowNode.
